requirements.txt
asgiref
crispy-bootstrap5
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
djangorestframework_simplejwt
Markdown
pillow
PyJWT
sqlparse
tzdata